Created a test suite for the warning modal (in generate) and fixed minor bug in the modal logic
